Improved the user interface
Former-commit-id: c4714435e8387188ea8f7ddb715da16b398a031c [formerly bde6de5d080e08f5b4ec5b568a67d5ec3f2195dd]
Former-commit-id: 55fa39384d0cabd203c4510263cf83c5d35470d7

+++ vaccine-management-center/src/main/java/pl/wysockif/optimizer/items/producers/Producer.java
@@ -3,11 +3,11 @@ package pl.wysockif.optimizer.items.producers;
 import java.util.Objects;
 
 public class Producer {
-    private final long id;
+    private final int id;
     private final String name;
     private final int dailyProduction;
 
-    public Producer(long id, String name, int dailyProduction) {
+    public Producer(int id, String name, int dailyProduction) {
         this.id = id;
         this.name = name;
         this.dailyProduction = dailyProduction;
@@ -26,7 +26,7 @@ public class Producer {
         return Objects.hash(id);
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
